Permisos solo para el creador Subir imágenes de las entradas

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Helpers\JwtAuth;
 use App\Post;
+use Illuminate\Contracts\Filesystem\FileNotFoundException;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
@@ -74,9 +75,7 @@ class PostController extends Controller
         if (empty($params_array)) return response()->json($data, $data['code']);
 
         // conseguir usuario identificado
-        $jwAuth = new JwtAuth();
-        $token = $request->header('Authorization', null);
-        $user = $jwAuth->checkToken($token, true);
+        $user = $this->getUser($request);
 
         // validar datos
         $validate = \Validator::make($params_array, [
@@ -112,6 +111,11 @@ class PostController extends Controller
 
     }
 
+    /**
+     * @param $id
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function update($id, Request $request) {
 
         // por dfecto devolvemos error generico
@@ -138,17 +142,35 @@ class PostController extends Controller
         if ($validate->fails()) {
             $data['message'] = 'Faltan datos';
             $data['errors'] = $validate->errors();
+
             return response()->json($data, $data['code']);
         }
+        // conseguir usuario identificado
+        $user = $this->getUser($request);
+//        if ($user->sub != $params_array['user_id']) {
+//            $data['message'] = 'Solo puedes actualizar post propios';
+//
+//            return response()->json($data, $data['code']);
+//        }
 
         // quitar lo que no queremos actualizar
         unset($params_array['id']);
         unset($params_array['user_id']);
         unset($params_array['created_at']);
+        unset($params_array['user']);
 
         // actualizar el registro
         // al actualizar con el metodo updateOrCreate devuelbe el objeto en si
-        $post = Post::where('id', $id)->updateOrCreate($params_array);
+        // upxate or create esta limitado a un filtro where se lo introducimod en un array
+        $where = [
+            'id'      => $id,
+            'user_id' => $user->sub
+        ];
+        /* $post = Post::where($where)->updateOrCreate($params_array);*/
+        $post = Post::where($where)->first();
+        if (empty($post) || !is_object($post)) return response()->json($data, $data['code']);
+
+        $post->update($params_array);
 
         $data = [
             'status'  => 'success',
@@ -169,28 +191,100 @@ class PostController extends Controller
      */
     public function destroy($id, Request $request) {
 
+        $post = null;
 
         // por dfecto devolvemos error generico
         $data = [
             'status'  => 'error',
             'code'    => 400,
             'message' => 'No se ha encontado post para borrar',
+            'post'    => $post
         ];
 
-        // conseguir registro
-        $post = Post::find($id);
+        // conseguir usuario identificado
+        $user = $this->getUser($request);
+
+        // conseguir registro solo para usuario autentificado
+        //$post = Post::find($id);
+        $post = Post::where('id', $id)
+            ->where('user_id', $user->sub)
+            ->first();
 
         if (!$post) return response()->json($data, $data['code']);
+
+        $data['message'] = ' solo puedes borrar tus posts';
+        if (!isset($post->id)) return response()->json($data, $data['code']);
+
 
         // Borrarlo
         $post->delete();
         // devolver algo
         $data = [
-            'status' => 'success',
-            'code'   => 200,
-            'post'   => $post
+            'status'  => 'success',
+            'code'    => 200,
+            'message' => 'Borrado corectamente',
+            'post'    => $post
         ];
         return response()->json($data, $data['code']);
 
     }
+
+    /**
+     * @param Request $request
+     * @return bool|object|null
+     */
+    private function getUser(Request $request) {
+        $jwAuth = new JwtAuth();
+        $token = $request->header('Authorization', null);
+        $user = $jwAuth->checkToken($token, true);
+
+        return $user;
+    }
+
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function upload(Request $request) {
+        $data = array(
+            'status'  => 'error',
+            'code'    => '400',
+            'message' => 'Error al subir la imagen',
+
+        );
+        // recoger datos peticionm
+        $image = $request->file('file0');
+
+        // validacion imagen
+
+        $validate = \Validator::make($request->all(), [
+            'file0' => 'required|image|mimes:jpg,jpeg,png,gif'
+        ], ['Tiene que ser una imagen valida']);
+
+        // guardar imagen
+        if (!$image || $validate->fails()) {
+            $data['errors'] = $validate->errors()->all();
+            return response()->json($data, $data['code']);
+        }
+
+        $image_name = time() . $image->getClientOriginalName();
+        // discos virtuales
+        try {
+            \Storage::disk('images')->put($image_name, \File::get($image));
+        } catch (FileNotFoundException $e) {
+            $data['errors'] = $e->getMessage();
+            return response()->json($data, $data['code']);
+        }
+
+        $data = array(
+            'status'  => 'success',
+            'code'    => '200',
+            'image'   => $image_name,
+            'message' => 'Imagen subida',
+        );
+
+        return response()->json($data, $data['code']);
+
+    }
+
 }
